Fix symbol 'after_speculation' is already defined
The original approach used inline asm to generate a label the linker
would then use. Unfortunately gcc decided to duplicate the block
containing the label, hence the error message.

Use a c label instead and store it's address to the global var at the
start of the function instead. This is a non-standard gcc extension but
it is more defined than using inline asm to insert a label.

Signed-off-by: Nicholas Miehlbradt <nicholas@linux.ibm.com>